Rename groupby and io files
